requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
twilio==8.12.0
python-dotenv==1.0.1
shopify==12.3.0
websockets==12.0
python-jose[cryptography]==3.3.0 